LayoutTests/imported/w3c/web-platform-tests/css/css-grid/grid-lanes/track-sizing/auto-repeat/intrinsic-auto-repeat/column-auto-repeat-auto-015-expected.txt


FAIL Auto-fit repeat tracks with auto size and auto item placement assert_equals: expected "250px 250px 250px 250px 0px 0px 0px 0px 0px 0px" but got "333.328125px 333.328125px 333.328125px"
